Fix double point

+++ src/main/java/com/mafia/statistics/MafiaStatisticsAPI/pyload/player/statistics/CoupleStatistics.java
@@ -19,5 +19,5 @@ public class CoupleStatistics {
 
     private Long wins;
 
-    private Double percentOfWins;
+    private Long percentOfWins;
 }
